1480. Running Sum of 1d Array
